Minor refactoring and typo fix in tests
