+++ src/main/scala/com/exasol/cloudetl/data/Row.scala
@@ -1,0 +1,3 @@
+package com.exasol.cloudetl.data
+
+final case class Row(val values: Seq[Any])
